#1779 should now work... infer() needs a rework

+++ nes-core/src/QueryCompiler/CodeGenerator/CCodeGenerator/TensorflowAdapter.cpp
@@ -60,10 +60,10 @@ void NES::TensorflowAdapter::infer(int n, ...){
     TfLiteTensor* input_tensor = TfLiteInterpreterGetInputTensor(interpreter, 0);
     int input_size = (int) (TfLiteTensorByteSize(input_tensor));
 
-    float* input = (float*) malloc(input_size);
+    int* input = (int*) malloc(input_size);
 
     for (int i = 0; i < n; ++i) {
-        input[i] = (float) va_arg(vl, double);
+        input[i] = (int) va_arg(vl, int);
     }
     va_end(vl);
